+++ src/jnsq.c
@@ -16,8 +16,8 @@
 #include <math.h>
 #include <stdint.h>
 
-void set_jnsq_bits(word_t* line, uint32_t inconsistency,
-				   const uint32_t n_attributes, const uint32_t n_words,
+void set_jnsq_bits(word_t* line, uint64_t inconsistency,
+				   const uint64_t n_attributes, const uint64_t n_words,
 				   const uint8_t n_bits_for_class)
 {
 	// How many attributes remain on last word with attributes
@@ -60,8 +60,8 @@ void set_jnsq_bits(word_t* line, uint32_t inconsistency,
 }
 
 void update_jnsq(word_t* to_update, const word_t* to_compare,
-				 uint32_t* inconsistency, const uint32_t n_attributes,
-				 const uint32_t n_words, const uint8_t n_bits_for_class)
+				 uint64_t* inconsistency, const uint64_t n_attributes,
+				 const uint64_t n_words, const uint8_t n_bits_for_class)
 {
 	// Set the line JNSQ
 	set_jnsq_bits(to_update, (*inconsistency), n_attributes, n_words,
@@ -82,7 +82,7 @@ void update_jnsq(word_t* to_update, const word_t* to_compare,
 /**
  * Adds the JNSQs attributes to the dataset
  */
-uint32_t add_jnsqs(dataset_t* dataset)
+uint64_t add_jnsqs(dataset_t* dataset)
 {
 	// Current line
 	word_t* current = dataset->data;
@@ -91,13 +91,13 @@ uint32_t add_jnsqs(dataset_t* dataset)
 	word_t* prev = current;
 
 	// Number of attributes
-	uint32_t n_attributes = dataset->n_attributes;
+	uint64_t n_attributes = dataset->n_attributes;
 
 	// Number of longs in a line
-	uint32_t n_words = dataset->n_words;
+	uint64_t n_words = dataset->n_words;
 
 	// Number of observations in the dataset
-	uint32_t n_observations = dataset->n_observations;
+	uint64_t n_observations = dataset->n_observations;
 
 	// Number of bits needed to store class
 	uint8_t n_bits_for_class = dataset->n_bits_for_class;
@@ -106,10 +106,10 @@ uint32_t add_jnsqs(dataset_t* dataset)
 	word_t* last = GET_LAST_LINE(dataset->data, n_observations, n_words);
 
 	// Inconsistency
-	uint32_t inconsistency = 0;
+	uint64_t inconsistency = 0;
 
 	// Max inconsistency found
-	uint32_t max_inconsistency = 0;
+	uint64_t max_inconsistency = 0;
 
 	// first line has jnsq=0
 	set_jnsq_bits(current, 0, n_attributes, n_words, n_bits_for_class);
